verification_scripts/requirements.txt
google-genai
python-dotenv
gTTS
